+++ just_audio_platform_interface/lib/just_audio_platform_interface.dart
@@ -103,6 +103,16 @@ abstract class AudioPlayerPlatform {
     throw UnimplementedError("setSpeed() has not been implemented.");
   }
 
+  /// Changes the pitch.
+  Future<SetPitchResponse> setPitch(SetPitchRequest request) {
+    throw UnimplementedError("setPitch() has not been implemented.");
+  }
+
+  /// sets skipSilence to true/false.
+  Future<SetSkipSilenceResponse> setSkipSilence(SetSkipSilenceRequest request) {
+    throw UnimplementedError("setSkipSilence() has not been implemented.");
+  }
+
   /// Sets the loop mode.
   Future<SetLoopModeResponse> setLoopMode(SetLoopModeRequest request) {
     throw UnimplementedError("setLoopMode() has not been implemented.");
@@ -126,6 +136,24 @@ abstract class AudioPlayerPlatform {
           SetAutomaticallyWaitsToMinimizeStallingRequest request) {
     throw UnimplementedError(
         "setAutomaticallyWaitsToMinimizeStalling() has not been implemented.");
+  }
+
+  /// On iOS and macOS, sets the
+  /// canUseNetworkResourcesForLiveStreamingWhilePaused option, and does nothing
+  /// on other platforms.
+  Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
+      setCanUseNetworkResourcesForLiveStreamingWhilePaused(
+          SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest request) {
+    throw UnimplementedError(
+        "setCanUseNetworkResourcesForLiveStreamingWhilePaused() has not been implemented.");
+  }
+
+  /// On iOS and macOS, sets the preferredPeakBitRate option, and does nothing
+  /// on other platforms.
+  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(
+      SetPreferredPeakBitRateRequest request) {
+    throw UnimplementedError(
+        "setPreferredPeakBitRate() has not been implemented.");
   }
 
   /// Seeks to the given index and position.
@@ -302,11 +330,13 @@ class IcyHeadersMessage {
 /// player instance.
 class InitRequest {
   final String id;
+  final AudioLoadConfigurationMessage? audioLoadConfiguration;
 
-  InitRequest({required this.id});
+  InitRequest({required this.id, this.audioLoadConfiguration});
 
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
         'id': id,
+        'audioLoadConfiguration': audioLoadConfiguration?.toMap(),
       };
 }
 
@@ -424,6 +454,44 @@ class SetSpeedResponse {
 }
 
 /// Information communicated to the platform implementation when setting the
+/// pitch.
+class SetPitchRequest {
+  final double pitch;
+
+  SetPitchRequest({required this.pitch});
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'pitch': pitch,
+      };
+}
+
+/// Information returned by the platform implementation after setting the
+/// pitch.
+class SetPitchResponse {
+  static SetPitchResponse fromMap(Map<dynamic, dynamic> map) =>
+      SetPitchResponse();
+}
+
+/// Information communicated to the platform implementation when setting the
+/// skipSilence.
+class SetSkipSilenceRequest {
+  final bool enabled;
+
+  SetSkipSilenceRequest({required this.enabled});
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'enabled': enabled,
+      };
+}
+
+/// Information returned by the platform implementation after setting the
+/// speed.
+class SetSkipSilenceResponse {
+  static SetSkipSilenceResponse fromMap(Map<dynamic, dynamic> map) =>
+      SetSkipSilenceResponse();
+}
+
+/// Information communicated to the platform implementation when setting the
 /// loop mode.
 class SetLoopModeRequest {
   final LoopModeMessage loopMode;
@@ -504,6 +572,46 @@ class SetAutomaticallyWaitsToMinimizeStallingResponse {
   static SetAutomaticallyWaitsToMinimizeStallingResponse fromMap(
           Map<dynamic, dynamic> map) =>
       SetAutomaticallyWaitsToMinimizeStallingResponse();
+}
+
+/// Information communicated to the platform implementation when setting the
+/// canUseNetworkResourcesForLiveStreamingWhilePaused option.
+class SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest {
+  final bool enabled;
+
+  SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest(
+      {required this.enabled});
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'enabled': enabled,
+      };
+}
+
+/// Information returned by the platform implementation after setting the
+/// canUseNetworkResourcesForLiveStreamingWhilePaused option.
+class SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse {
+  static SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse fromMap(
+          Map<dynamic, dynamic> map) =>
+      SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse();
+}
+
+/// Information communicated to the platform implementation when setting the
+/// automaticallyWaitsToMinimizeStalling option.
+class SetPreferredPeakBitRateRequest {
+  final double bitRate;
+
+  SetPreferredPeakBitRateRequest({required this.bitRate});
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'bitRate': bitRate,
+      };
+}
+
+/// Information returned by the platform implementation after setting the
+/// automaticallyWaitsToMinimizeStalling option.
+class SetPreferredPeakBitRateResponse {
+  static SetPreferredPeakBitRateResponse fromMap(Map<dynamic, dynamic> map) =>
+      SetPreferredPeakBitRateResponse();
 }
 
 /// Information communicated to the platform implementation when seeking to a
@@ -725,6 +833,166 @@ class VisualizerFftCaptureMessage {
     required this.samplingRate,
     required this.data,
   });
+}
+
+/// Information communicated to the platform implementation when setting the
+/// audio load configuration options.
+class AudioLoadConfigurationMessage {
+  final DarwinLoadControlMessage? darwinLoadControl;
+  final AndroidLoadControlMessage? androidLoadControl;
+  final AndroidLivePlaybackSpeedControlMessage? androidLivePlaybackSpeedControl;
+
+  const AudioLoadConfigurationMessage({
+    required this.darwinLoadControl,
+    required this.androidLoadControl,
+    required this.androidLivePlaybackSpeedControl,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'darwinLoadControl': darwinLoadControl?.toMap(),
+        'androidLoadControl': androidLoadControl?.toMap(),
+        'androidLivePlaybackSpeedControl':
+            androidLivePlaybackSpeedControl?.toMap(),
+      };
+}
+
+class DarwinLoadControlMessage {
+  /// (iOS/macOS) Whether the player will wait for sufficient data to be
+  /// buffered before starting playback to avoid the likelihood of stalling.
+  final bool automaticallyWaitsToMinimizeStalling;
+
+  /// (iOS/macOS) The duration of audio that should be buffered ahead of the
+  /// current position. If not set or `null`, the system will try to set an
+  /// appropriate buffer duration.
+  final Duration? preferredForwardBufferDuration;
+
+  /// (iOS/macOS) Whether the player can continue downloading while paused to
+  /// keep the state up to date with the live stream.
+  final bool canUseNetworkResourcesForLiveStreamingWhilePaused;
+
+  /// (iOS/macOS) If specified, limits the download bandwidth in bits per
+  /// second.
+  final double? preferredPeakBitRate;
+
+  DarwinLoadControlMessage({
+    required this.automaticallyWaitsToMinimizeStalling,
+    required this.preferredForwardBufferDuration,
+    required this.canUseNetworkResourcesForLiveStreamingWhilePaused,
+    required this.preferredPeakBitRate,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'automaticallyWaitsToMinimizeStalling':
+            automaticallyWaitsToMinimizeStalling,
+        'preferredForwardBufferDuration':
+            preferredForwardBufferDuration?.inMicroseconds,
+        'canUseNetworkResourcesForLiveStreamingWhilePaused':
+            canUseNetworkResourcesForLiveStreamingWhilePaused,
+        'preferredPeakBitRate': preferredPeakBitRate,
+      };
+}
+
+class AndroidLoadControlMessage {
+  /// (Android) The minimum duration of audio that should be buffered ahead of
+  /// the current position.
+  final Duration minBufferDuration;
+
+  /// (Android) The maximum duration of audio that should be buffered ahead of
+  /// the current position.
+  final Duration maxBufferDuration;
+
+  /// (Android) The duration of audio that must be buffered before starting
+  /// playback after a user action.
+  final Duration bufferForPlaybackDuration;
+
+  /// (Android) The duration of audio that must be buffered before starting
+  /// playback after a buffer depletion.
+  final Duration bufferForPlaybackAfterRebufferDuration;
+
+  /// (Android) The target buffer size in bytes.
+  final int? targetBufferBytes;
+
+  /// (Android) Whether to prioritize buffer time constraints over buffer size
+  /// constraints.
+  final bool prioritizeTimeOverSizeThresholds;
+
+  /// (Android) The back buffer duration.
+  final Duration backBufferDuration;
+
+  AndroidLoadControlMessage({
+    required this.minBufferDuration,
+    required this.maxBufferDuration,
+    required this.bufferForPlaybackDuration,
+    required this.bufferForPlaybackAfterRebufferDuration,
+    required this.targetBufferBytes,
+    required this.prioritizeTimeOverSizeThresholds,
+    required this.backBufferDuration,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'minBufferDuration': minBufferDuration.inMicroseconds,
+        'maxBufferDuration': maxBufferDuration.inMicroseconds,
+        'bufferForPlaybackDuration': bufferForPlaybackDuration.inMicroseconds,
+        'bufferForPlaybackAfterRebufferDuration':
+            bufferForPlaybackAfterRebufferDuration.inMicroseconds,
+        'targetBufferBytes': targetBufferBytes,
+        'prioritizeTimeOverSizeThresholds': prioritizeTimeOverSizeThresholds,
+        'backBufferDuration': backBufferDuration.inMicroseconds,
+      };
+}
+
+class AndroidLivePlaybackSpeedControlMessage {
+  /// (Android) The minimum playback speed to use when adjusting playback speed
+  /// to approach the target live offset, if none is defined by the media.
+  final double fallbackMinPlaybackSpeed;
+
+  /// (Android) The maximum playback speed to use when adjusting playback speed
+  /// to approach the target live offset, if none is defined by the media.
+  final double fallbackMaxPlaybackSpeed;
+
+  /// (Android) The minimum interval between playback speed changes on a live
+  /// stream.
+  final Duration minUpdateInterval;
+
+  /// (Android) The proportional control factor used to adjust playback speed on
+  /// a live stream. The adjusted speed is calculated as: `1.0 +
+  /// proportionalControlFactor * (currentLiveOffsetSec - targetLiveOffsetSec)`.
+  final double proportionalControlFactor;
+
+  /// (Android) The maximum difference between the current live offset and the
+  /// target live offset within which the speed 1.0 is used.
+  final Duration maxLiveOffsetErrorForUnitSpeed;
+
+  /// (Android) The increment applied to the target live offset whenever the
+  /// player rebuffers.
+  final Duration targetLiveOffsetIncrementOnRebuffer;
+
+  /// (Android) The factor for smoothing the minimum possible live offset
+  /// achievable during playback.
+  final double minPossibleLiveOffsetSmoothingFactor;
+
+  AndroidLivePlaybackSpeedControlMessage({
+    required this.fallbackMinPlaybackSpeed,
+    required this.fallbackMaxPlaybackSpeed,
+    required this.minUpdateInterval,
+    required this.proportionalControlFactor,
+    required this.maxLiveOffsetErrorForUnitSpeed,
+    required this.targetLiveOffsetIncrementOnRebuffer,
+    required this.minPossibleLiveOffsetSmoothingFactor,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'fallbackMinPlaybackSpeed': fallbackMinPlaybackSpeed,
+        'fallbackMaxPlaybackSpeed': fallbackMaxPlaybackSpeed,
+        'minUpdateInterval': minUpdateInterval.inMicroseconds,
+        'proportionalControlFactor': proportionalControlFactor,
+        'maxLiveOffsetErrorForUnitSpeed':
+            maxLiveOffsetErrorForUnitSpeed.inMicroseconds,
+        'targetLiveOffsetIncrementOnRebuffer':
+            targetLiveOffsetIncrementOnRebuffer.inMicroseconds,
+        'minPossibleLiveOffsetSmoothingFactor':
+            minPossibleLiveOffsetSmoothingFactor,
+      };
 }
 
 /// Information about an audio source to be communicated with the platform